requirements/requirements_prediction.txt
torch
torchvision
flask
pillow
flask_cors
opencv-python
numpy